training/requirements.txt
# ML Training Pipeline Requirements
tensorflow>=2.13.0
tensorflowjs>=4.0.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
albumentations>=1.3.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
seaborn>=0.12.0